Create BAGofMARBLES
This Bag of Marbles ask for user to enter a list of numbers in which will then pull from the list the number on the marble you want to remove then will display the numbers of which marble is left in the bag. The program will also display whether the number is in the bag or not.